+++ elements/simple-colors/demo/index.html
@@ -8,6 +8,7 @@
     <script type="module">
       import '../node_modules/@polymer/iron-demo-helpers/demo-snippet.js';
       import '../simple-colors.js';
+      import './simple-colors-demo.js';
     </script>
     <style is="custom-style" include="demo-pages-shared-styles"></style>
   </head>
@@ -16,9 +17,16 @@
       <h2>simple-colors</h2>
       <demo-snippet>
         <template>
-          <simple-colors>Test</simple-colors>
+          <style is="custom-style" include="simple-colors">
+            p {
+              background-color: var(--simple-colors-default-theme-red-1);
+            }
+          </style>
+          <p>Test</p>
+          <simple-colors><p>Test</p></simple-colors>
         </template>
       </demo-snippet>
+      <p>See <a href="colors.html">Colors</a></p>
     </div>
   </body>
 </html>
