+++ elements/simple-colors/index.html
@@ -8,7 +8,7 @@
   <script type="module" src="node_modules/@polymer/iron-component-page/iron-component-page.js"></script>
 </head>
 <body>
-  <a href="demo/index.html">demo</a>
-  <!--iron-component-page></iron-component-page-->
+  <!--a href="demo/index.html">demo</a-->
+  <iron-component-page></iron-component-page>
 </body>
 </html>
